main: use portable uint64_t for cred_ll_read bounds check
- Compare off and node->len using uint64_t instead of size_t.
- This prevents potential 64-bit to 32-bit silent truncation of signed read offsets on 32-bit platforms, avoiding out-of-bounds reads.

+++ src/main.c
@@ -416,7 +416,7 @@ static void cred_ll_read(fuse_req_t req, fuse_ino_t ino __attribute__((unused)),
         return;
     }
 
-    if ((size_t)off >= node->len) {
+    if ((uint64_t)off >= (uint64_t)node->len) {
         fuse_reply_buf(req, NULL, 0);
         return;
     }
